refactor: improve code
